+++ src/kernel/utilities/shell/shell.c
@@ -184,7 +184,7 @@ void launch_shell(int n) {
       }
     } else if (strEql(ch, "lspci")) {
       printf("\n");
-      for (uint8_t bus = 0; bus < PCI_MAX_BUSES; bus++) {
+      for (uint16_t bus = 0; bus < PCI_MAX_BUSES; bus++) {
         for (uint8_t slot = 0; slot < PCI_MAX_DEVICES; slot++) {
           for (uint8_t function = 0; function < PCI_MAX_FUNCTIONS; function++) {
             if (!FilterDevice(bus, slot, function))
